Add --host flag to set the server bind address
This adds a `--host` flag so that you can specify where the jobly server will bind.

Before this change, it was hard-coded to 0.0.0.0 which binds to every available interface. For security, I only want it to listen on localhost. With this change it can be done with: `jobly server --host localhost`:

```
$ bundle exec bin/jobly server --host localhost
Starting server
Puma starting in single mode...
* Puma version: 6.3.0 (ruby 3.2.0-p0) ("Mugi No Toki Itaru")
*  Min threads: 0
*  Max threads: 5
*  Environment: development
*          PID: 36472
* Listening on http://127.0.0.1:3000
* Listening on http://[::1]:3000
Use Ctrl-C to stop
...
```

I've matched the flag names to what `rackup` uses.